requirements.txt
pytorch
numpy
nltk
tqdm
pandas
jieba
tensorflow
tensorboardX
cython
regex
matplotlib
pyyaml
sacremoses
editdistance
soundfile
sentencepiece
pyarrow
python_speech_features
torchaudio
hydra
sacrebleu==1.5.1
hydra-core==1.0.0
omegaconf==2.0.5